Global refactor of everything
ScriptAction enum is gone, replaced by specialized "change" classes :
 - AddCommand
 - RemoveCommand
 - MoveCommand
 - SetCommand

We still use two containers to represent sections : ScriptSection and
ConfigSection, although they now are now not related. ScriptSection may
contain command objects, while ConfigSection only contains configuration
entities, represented by :
 - ConfigItem (wraps an OrderedDict)
 - ConfigSetting (wraps an optional string and an OrderedDict)

The command classes are able to apply their changes to ConfigSection
objects, which in turns makes MikrotikConfig able to apply a
MikrotikScript. This effectively reverses the original feature, and
allows to apply a configuration script to a configuration object.

This allowed to add a unit test that checks that the changes do result
in the target configuration when passed back into the parser, which is
now able to parse any configuration command.

About parsing : we use a mix of lexing/matching to parse the text as
configuration. Lines are split using a dedicated line lexer that handles
quoted newlines, commands and their parameters are matched using regular
expressions, and key-value pairs are matched using a split shell
splitter.

This file MAY be a good example of advanced text parsing, but I don't
think I did it really correctly.